# Speed up install of environments with dev packages