@bernhardkaindl 2021-11-08
### Initial analysis of the codecov/patch code coverage issue:
@tmadlener  From looking at the report diff, the red background color of the line number and the hover text of the 🛈 symbol left to the red line number indicates that Line 1420 is not covered by tests: https://app.codecov.io/gh/spack/spack/compare/27167/diff#D1L1420
Adding a test to run with `spack unit-test` would be needed to make it pass.
Since the new line is in `def _get_overwrite_specs()` I'm assuming a test which uses `spack install --overwrite` should cover it, and there are appear to be some, according to `grep -r overwrite lib/spack/spack/test`. Maybe you have an idea why none of them appears to cover that line (assuming everything else I see is correct).