@becker33 2021-11-08
If we're not able to figure out why the codecov is dropping, another option would be to avoid memoizing and just precompute the set of all specs we need to check. Something like:

```
def _get_overwrite_specs(self):
    specs_to_check = set()
    for dag_hash in self.concretized_order:
        root_spec = self.specs_by_hash[dag_hash]
        specs_to_check.update(root_spec.traverse(root=True))
    return [s for s in specs_to_check if self._spec_needs_overwrite(s)]
```

Since specs are appropriately hashable, that should eliminate the duplicate work that memoizing eliminated.