@tmadlener 2021-11-08
> If we're not able to figure out why the codecov is dropping, another option would be to avoid memoizing and just precompute the set of all specs we need to check.

That sounds like an overall nicer solution in any case in my opinion. I will give that a go and see if it works.