@tmadlener 2021-11-09
Ah, I just realized why I went with the memoization. `_spec_needs_overwrite` calls itself as well to figure out if dependent packages need to be rebuilt because (dev) dependencies have changed. Only having the set that collects all specs upfront doesn't help because only the first layer of dependency tree walking is removed, but below that nothing changes.

I have not been able yet to come up with some clever changes to `_spec_needs_overwrite` yet, that would work without the memoization and still respect the dependencies.