@bernhardkaindl 2021-11-28
@tmadlener, the CI Pipeline failure seems to be from:
> Error: Pipeline generation failed due to the presence of the following specs that are known to be broken in develop:
> raja@0.11.0%gcc@7.5.0~cuda+examples+exercises~ipo+openmp~rocm+shared~tests

`raja` might have been fixed by: https://github.com/spack/spack/commit/6f8fd5b7a7096a1eee29988969197c91f7ba889f

Just to be one the safe side: Maybe could to rebase this PR to the current develop? (Just rerunning the CI could work too, but maybe it's a good idea to also rebase it)